util: Add a string reference type to avoid hardcoded string lengths.
Because doing this gets silly pretty quickly, especially if you make a
mistake. ("Name2: " is 7 characters long, not 6.) By using C99 compound
literals behind a macro, we can nicely retain the visual syntax-
highlighted structure of the code.

This covers everything outside of libthtk, which doesn't use the util
library.