thanm: Spell out the zlib-ng source files, ignoring gz* in the process.
CMake recommends GLOB to not be used for source files. Since we don't
use zlib-ng's CMakeLists.txt (probably to avoid the needless compilation
of its example code?), we'll always have to manually copy any of their
changes anyway.

Also, this is the quickest way to make the latest zlib-ng code compile
*without* having to care about errors. :P