thecl: Look up ≥th10 instruction formats even if they have no parameters.
Because as seen in 06269fb, not showing that error when dumping an
unknown instruction only defers it to the first time someone wants to
recompile the file containing it. This typically causes great confusion
when trying to do Skipgame on Day 1 of a new game release. ("Wait, but
all files dumped fine, what's going on there?!")

Funded by zorg.