thecl: add a few unused formats
ins_18, 19 and 20 do something related to subs called with ins_16: first parameter of all of them is the ID used in ins_16. ins_18 and 19 set some boolean value in a structure acquired based on the ID to true and false respectively, while ins_20 writes the value given in the second parameter to the same struct (not the same value as 18/19). That said, I still have no idea what they actually do, I couldn't find any instances of the game checking their values at all...

ins_24 is same as ins_23, except it takes a float as the waiting time. What? A frame is a frame and you can't say it's only a half? Well, ZUN would like to have a word with you