thanm: variables and expressions
Still incomplete, but expressions for v8 should fully work (though some things like NEG are still missing).
--------
It's now possible to declare variables like in thecl, and the compiler will automatically assign registers to them (when you later try to use the register that was assigned to a variable, you'll get a warning). You can also write expressions like in thecl, and the compiler will output instructions that evaluate the expression (and use some temporary registers if necessary).