thecl: fix inline sub instr deletion issues
The first bug was that the labels were not adjusted to account for instructions that weren't copied. The second bug was that not unsetting the return value flag after copying it into another inline sub would cause these instructions to still be treated as "return value outputters" even though they are not.