thanm: UM full ver support
Has some issues because ins_439 in VD has a different format.

VD photo.anm now throws a bunch of warnings when recompiling, but it still works. It's the only workaround I could come up with, ideally in the future the game number gets passed as a command line parameter so that the proper ins_439 format can be used.