Update extlibs
zlib-ng moved to a new repo, and needed a couple of changes to the
CMakeLists in order to make it work.

Use the following to update remotes:
    git submodule sync --recursive

Also, recently I've switched libpng from the slow, but official sf repo
to the author's personal repo. Well, uh, he's been dead since 2018. The
current maintainer pushes to both repos tho, so it's fine. Also, this
means it took me 4 years to figure out how to change git submodule url.