thanm: bring back image composition
Turns out #111 wasn't that much of an issue after all.

This is a partial revert of 58be1106561eef40a08fceeedb93ba932d9f6db5, in
a sense that I reimplemented the stuff that I removed in that commit.

Now we have both combined and separate (-u) extraction implemented in a
somewhat elegant manner. Every entry now has a link to the next entry of
the same name. For separate extraction, we simply don't build the links.

Additionally, th19 is now fully supported in the combined mode (except
for replacement and for JPEGs). thanm will opportunistically try to
avoid recompressing PNGs. The -uu flag creates all the necessary
conditions for that.