thecl10: introduce type 'H' for th20 ins_46/ins_47
ZUN's compiler, it seems, treats these arguments as 'D', but the vm
behaves as if these were SS/Sf when looking up the bits in the
param_mask. Vanilla game never uses variables in those arguments, but
mods may wish to do so.

The new 'H' type is equivalent to 'D', but a bit in the param_mask is
skipped when reading/writing the instruction. param_count is not
affected.